requirements.txt
langchain
openai==v0.28.1
huggingface_hub
python-dotenv
streamlit